Breast Cancer Classification - Bin into Malignant/Benign (97.4% accuracy)
The data set for this problem was obtained from Kaggle as a .csv file. This program utilizes Logistic Regression, Decision Trees, and KNN to accurately classify 569 women that have been pre-diagnosed as malignant/benign. 2 models for each classification type, with and without pre-processing, have been created to better compare the results.